refactor(analytics): extract ConversionSection component (Phase52f conversion analysis)
